backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
python-dotenv==1.0.0
pydantic==2.5.3
requests==2.31.0
pdfplumber==0.10.3
openpyxl==3.1.2
python-docx==1.1.0
reportlab==4.0.7
sqlparse==0.4.4
pandas==2.0.3
